Replace sliding panel with scrollview
This allows
1) Sliding of the drawer to happen in the theme viewpager
2) Seamless sliding from viewpager all the way to the bottom most
theme component checkbox (sounds).

Before this patch the UX was a little awkward since the theme
components were in their own scrollview. A user could not scroll down at all
in the theme preview viewpager and would have to use two swipes to
get to the very bottom.

Change-Id: I1a0a30edfdd29d56db0aa40d06114f8bbaa39331