Vertical center Themes by using a viewpager
Transitioning from a center collapsed view to an expanded view
requires a layout change. Animating children during a layout change
requires the children to be placed in a grandparent's ViewOverlay
so that they do not get clipped. A ListView cannot remove a child view,
so ViewOverlay cannot be used with ListView items, therefore a ViewPager
was used instead.

Change-Id: I665b89bf85d7466334db8a64bf37475e9e268efc