Destroy the LOADER_ID_APPLIED in onLoadFinished.
This loader was being called several times during, and immediately
after, a theme change causing the items in the additional cards to
revert back to their previous setting.

i.e. If you change the boot animation, apply the change, and then
expand the card you would see the boot animation was back to the
previous one even though the selector indicates that the current one
is the one you just applied.

Change-Id: I1125d9d1a7a82ef0c6f2b619e4fb5f02b0efc30e