Catch RuntimeException when attempting to draw a recycled bitmap.
It is possible for the boot animation to be in the middle of a call
to onDraw when a new animation is being loaded.  This can result in
one of the bitmap buffers being recycled which will throw a
RuntimeException when trying to draw it to the canvas.  Worse case
we lose a frame or two of the animation when this occurs.

Change-Id: Ied2fb77b787f4a4d6141780113ebac2a177f6e57