Remove the final RelativeLayout from fragment_pager_list
With the switch to a LinearLayout we must give the additional cards
a negative margin.  This allows them to be positioned below the
navbar card.

This is done because previously the title card was being set to
View.GONE and this caused the additional cards to abruptly move up
when the title card changes to GONE.  By setting it to View.INVISIBLE
and using a negative margin we can get a smooth animation and not lose
any functionality in the process.

Change-Id: Ifbbf67f3e301a26fa3d30ef79a3a07416fba5837