Scroll the ScrollView to reveal content behing component selector.
The current implementation shifted the entire content view up which
caused the top to be off screen and the user unable to scroll all the
way to the top while the component selector was out.  This patch
animates scrolling the content in the ScrollView and adds a space view
to the bottom so that all content will be visible above the component
selector.

Change-Id: I48577e941b6429db3673b6769f44f3519097b2e8