Show live wallpaper in MyThemeFragment
If the user is currently using a live wallpaper, we can view it
by adding a SurfaceView on top of the wallpaper ImageView.  Since
the surface view is z ordered so that it is behind the window
holding the surface view it acts as a window to the current wallpaper.

Change-Id: I10a2507bef125119ff088e96acb726f6ca84c6ae