Use NavBarSpace for padding the bottom of some layouts
We've used Space in enough layouts that it makes sense to create
a view that can take care of all the logic regarding padding the
bottom of layouts so content goes above the navigation bar.  This
view simply sets the height to 0 in onMeasure() when no navigation
bar is detected.  It leaves the height to whatever the super class
had set it to when there is a navigation bar.

Change-Id: I908c6a8a01ba4f761335909e9ecfb48a8a3007d2