Handle theme changes better
If the user closes out the app during a theme change it is possible
to end up in a state where the chooser incorrectly displays the applied
theme as the one that was perviously applied before the change.  When
in this state it is also possible for the theme change listener to not
be triggered and the card displays an empty progress bar.

Change-Id: I323cfdd2c9c6d1fd9d50ba7f6f1bae7e355c0f62