Don't add unsupported components to mSelectedComponentsMap
We add a package to the supported components map when the various
loadXXXXXX() methods are called but this is not correct.  The theme
may not support that component and in that case we should not add
the package to the supported map.  If we do the theme service attempts
to apply those components and in the case of fonts, you end up with
all text gone.

Change-Id: Icd718f4f69f8296a9b5c11f49f5e169373c8b6f6