Delete packages from inside app
Deleting themes offers a bad user experience in its current form.
Since PackageInstaller does not let us know that the package
removal was successful, the theme will remain in the viewpager until
we finally receive the broadcast from the PackageManager that the
theme was removed.

This patch adds the DELETE_PACKAGES permission and deletes the package
without using the PackageInstaller.  This allows us to know immediately
that the package was removed and we can update the themes right away.

Change-Id: I229f2fbadc6f614a8180fb04735906f86e5e96fa