Optimize adding, removing, and re-ordering fragments
FragmentStatePagerAdapter saves the state of the fragments based on
their position.  This works great until you start to add and remove
fragments.  The fragments get their state restored incorrectly and
we end up with duplicates or gaps.  We initially ended up recreating
the pager adapter whenever we needed to add or remove a theme to
get by this limitation.  This patch brings in NewFragmentStatePagerAdapter
found at https://code.google.com/p/android/issues/detail?id=37990 to
provide a better work around for this bug.

We still re-create the pager adapter if we change a theme since the
MyThemeFragment requires some special attention.

Change-Id: Ibaa6854085f8653a6eb6dbe288bdbed871551ed2