Add handling of themes that are being processed
When a theme is still being processed by the ThemeService we show
a "processing" overlay to let the user know their theme is not
quite ready to be applied.  Once the theme is done processing the
overlay fades out and the user can apply the theme.

We also do not show the "theme installed" notification for a theme
that is being processed until it is finished processing.

Change-Id: I0486da3a5e2d0b55c2b3828613ace7e2ccf460a2