Destroy loader when done loading components
A strange bug was uncovered where items keep loading into the
selector which causes many duplicates.  Once we load the data
we can simply destroy the loader so we will not get any further
updates.

The root cause appears to be caused by the post processing of
themes in the ThemeService.  The ThemesProvider listens for changes
to theme resources and updates the preview blobs which triggers the
loaders to be triggered.

Change-Id: Ifbeddadd2cd47d0f04755036fbb432d0fafb3df2