Wait for fragment to collapse before restarting a loader
Bad things happen when a theme is expanded and a new theme is
installed and added to the view pager.  This patch sets a flag
that lets the ChooserActivity know it should restart the loader
once the fragment is collapsed.  Once the fragment is collapsed
the loader will be restarted and new themes will be added to the
view pager.

Change-Id: I2b5fe64782ce278b7e8b09e4ff24755faa080bec