MyThemeFragment should only be applying changed components
The current logic sends all the comonents to the theme service,
even when only one or two components changed in the MyThemeFragment.
This patch will create a map of only the components that changed
and pass those onto the ThemeService.

This patch also fixes the issue with fillMissingComponentsWithDefault
which should only be used for applying full themes.

Change-Id: Id47e771e58afd6d2f5065820617b0c56cecc2405