Revert "Fix theme installed notification not showing up"
This reverts commit a4d0390b36adc5089b1819712de7c6d7b78ac54f.
We ned to keep track of themes that are installed and being processed
by the theme service.  Themes are processed on boot and can cause false
positives if we assume receiving this broadcast constitutes a theme
being installed.

The root cause for not showing install notifications was due to the
Set being null when the preference for themes being processed does not
exist.  We correct this by create a new Set if it is null.

Change-Id: I1e762abb6248b282f076a1cf014fc7864f6ce0a3