Add better error handling when loading audibles
We were checking if the ringtoneUri was null but not if the actual
Ringtone returned for that uri was null.  This patch handles that
case and simply shows the "Add ...." so the user can add a valid
audible.

Change-Id: I4c2320c6ed7ccd825ffcec411bdffc858b08741e
REF:TOMATO-186
(cherry picked from commit 77a2e801640bff65cf7b0d9fadb2364b9531579d)