Send broadcast when theme chooser opened and theme removed
This introduces two new actions that will be broadcasted to the
store, or any app that is granted the com.cyngn.themes.permission.THEMES_APP
permission, which requires the same signature as the Theme Chooser.

If a theme is uninstalled from the chooser, a broadcast is sent along with
a String extra with the key "package" that identifies the package name
of the theme being removed.

Change-Id: I6b3f4e19ebb1b8d763d7e0038591fccbbe1809f2