Re-use existing typeface when setting to bold
We set the text for the "default" entry in the component selector
to bold but pass in null for the typeface.  This ends up using the
DEFAULT_BOLD typeface rather than lato and can result in a native
crash in libminikin.  There is still an issue with fonts at the
native layer but this patch helps mitigate the issue from crashing
the chooser.

Change-Id: I9e4cde2452c18d538db5d27ea2acf8c92ff27aad
REF: THEMES-537