Remove deprecated theme showcase for shop themes
Now that theme showcase is deprecated and the store is shipping on
our devices, we no longer need the string for themes showcase and
the code that relied on that.  This patch adds a check for the
theme store, even though it SHOULD be installed, and hides the
shop themes button in this case.

Change-Id: I89746914698305eb736c354440218d37dcdce02c