AppThemer: System theme always has overlay for app
Whenever using app themer to switch an app to the system theme, the
user is presented with a message stating that this theme does not
explicitly overlay the current app.  System theme is really an
absence of a theme so we should not show this message when applying
the system theme to an app.

Change-Id: I4087744c325dcd07abffe9a893ecc8fa4c9f1e07