Register/Unregister observer on window visibility change
If the ThemeChooser is in a paused state, the ThemeObserver in the
component selector will still fire off when a theme is added or
removed.  This causes an NPE when the loader is restarted.

This patch unregisters the content observer when the window is no
longer visible and then re-registers it once the window is visible
again.

Change-Id: I628b269eb95a7fb521470ff3b33845086ee30461
TICKET: CHOOSER-108