Use Lato font on Animated Lock Screen thumbnail text
To be consistent with the rest of the text views in the chooser,
we need to use Lato font on the animated lock screen thumbnail
(rather than Roboto as initially requested).

In order to reuse as much code as possible, LatoTextView is now
extending FittedTextView. A new attribute 'autoFitText' is
introduced to instruct whether fitted text is desired.

Change-Id: I3209fe7f06177bc4025cbf964d0c2dd79325b9c0
TICKET: CHOOSER-113