Fix Lock Screen card scroll animation
The lock screen card is being pushed to the top of the screen
during the scroll animation and then slid down (to make sure
the whole card is visible) if the device has hardware buttons.

This patch will make sure the UX is the same across all devices.
The card will be positioned in the middle of the screen once the
animation is complete.

Change-Id: Ic24c638dfa90f5f58424bc931eb5235befd31880
TICKET: CHOOSER-114