Show the apply icon when a theme component is changed
To improve the ThemeChooser UX, the apply icon will now be displayed
when the user makes a change in the component selector, rather than
waiting until user presses the back button. The icon will remain
visible even if the user jumps between components cards, as long as
there's at least one change to apply. The icon will be dismissed
if the user reverts the changes by selecting the currently applied
components.

Change-Id: I3cdde376a89d4456a2f68ca7957cc8e1db7c27e9
TICKET: CYNGNOS-2077