Check for system alert window permission
It is possible that the user disables theme chooser from being able
to draw on top of other applications.  If this happens theme chooser
will crash due to a security exception.  This patch checks that we
still have the permission and if not, requests it from the user.

Change-Id: Ic4fc12293d4fa2c544d2e54ed8279dc07d7493f6
TICKET: CYNGNOS-3141